+++ CHANGELOG.md
@@ -11,6 +11,7 @@ All notable changes to this project will be documented in this file.
 - HASPmota rounds to nearest int values passed as 'real' (#21599)
 - Berry automatic rounding of float to int when calling C mapped functions (#21601)
 - Berry add `math.round` (#21602)
+- Support for QMP6988 temperature and pressure sensor
 
 ### Breaking Changed